==> PER/2019/1.D6/Rcodes/7b.ECM.R <==
#########################################################
# Proyecto MRP - Leave no one behind                    #
# Mapas de pobreza CEPAL                                #
# Lectura y preparación de las bases de datos           #
# Autor: Stalyn Guerrero & Andrés Gutiérrez             #
#########################################################
rm(list =ls())

cat("\f")
# Loading required libraries ----------------------------------------------

library(rstan)
library(rstanarm)
library(data.table)
library(dplyr)
library(magrittr)
library(forcats)
library(tidyr)
library(reshape2)
library(stringr)
library(ggplot2)
library(scales)
library(bayesplot)
library(gridExtra)
library(ggalt)
library(usmap)
library(gridExtra)
library(scales)
library(kableExtra)
library(formatR)
library(patchwork)

theme_set(bayesplot::theme_default())
library(tidyverse)

source("0Funciones/funciones_mrp.R", encoding = "UTF-8")

# Loading data ------------------------------------------------------------

encuesta_mrp <- readRDS("PER/2019/1.D6/Data/encuesta_mrp.rds")
poststrat_df <- readRDS("PER/2019/1.D6/Data/poststrat_df.RDS") %>% 
  filter(anoest != "99")
tasa_desocupados <- readRDS("PER/2019/1.D6/Data/tasa_desocupacion.rds")
fit <- readRDS("PER/2019/1.D6/Data/fit_bayes.rds")

# Poststratification at the National Level --------------------------------

byAgrega <-
  grep(
    pattern =  "^(n|X|F|pobreza|tasa_desocupacion|epred_mat|gk|mpio|lp|stable_lights|crops.coverfraction|urban.coverfraction)",
    x = names(poststrat_df),
    invert = TRUE,
    value = TRUE
  )

byAgrega

# Posterior_epred returns the posterior estimates for the different subgroups stored in the
# poststrat_df dataframe.

epred_mat <- posterior_epred(fit, newdata = poststrat_df, type = "responde")

## validacion de los valores posteriores
summary(rowMeans(epred_mat))
summary(colMeans(epred_mat))

summary(as.numeric(epred_mat))
hist(as.numeric(epred_mat))

length(epred_mat[which(epred_mat == 0)])

prop_gzero <- function(x) mean(x < 0)
prop_gzero(epred_mat)

epred_mat[which(epred_mat < 0)] = 0

# Resultados nacionales ---------------------------------------------------
(mrp_estimate_Ingresolp <-
  Aux_Agregado(poststrat = poststrat_df,
             epredmat = epred_mat,
             byMap = NULL)
)
# Poststratification at the State level -----------------------------------

byAgrega <-
  grep(
    pattern = "mpio",
    x = byAgrega,
    value = TRUE,
    invert = TRUE
  )

byAgrega <- t(combn(byAgrega, 2))
byAgrega <- rbind(c("mpio","mpio" ), byAgrega)
# resultados para ingreso medio

mrp_estimate = map(1:nrow(byAgrega),
             ~poststrat_df %>% 
               group_by_at(vars("mpio", byAgrega[.x,])) %>%
               summarise(Benchmarking_estimate = sum(n * pobreza2) / sum(n), .groups = "drop") %>% 
               ungroup())


mrp_ECM <- map(1:nrow(byAgrega), function(i) {
     Aux_Agregado(poststrat_df,
                 epredmat = epred_mat,
                 byMap = c("mpio", byAgrega[i, ])) 

  })

tablas <- map2(mrp_estimate, mrp_ECM, inner_join) %>% 
  map(~.x %>%
        mutate(mrp_cv = mrp_estimate_se/Benchmarking_estimate*100))

nom_tabs <- c("mpio", apply(byAgrega[-1,],MARGIN = 1,paste0, collapse = "_"))
names(tablas) <- nom_tabs


tablas %>% map_df(~.x %>% summarise(max = max(mrp_cv)), .id = "agregado")
tablas %>% map(~.x %>% filter(mrp_cv > 50))


openxlsx::write.xlsx(tablas, file = "PER/2019/1.D6/Output/tablas_ECM.xlsx", 
                     overwrite = TRUE)
